+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 * `Reduce`
 * `Overlap`
 
+### Changed
+* `Contains` and `Missing` are unary
+
 ## [0.4.0](https://github.com/coady/iterset/releases/tag/v0.4.0) - 2025-04-26
 ### Added
 * `Equal` and `EqualCounts`